+++ tests/test_modified_cosine.py
@@ -18,7 +18,7 @@ def test_modified_cosine_without_precursor_mz():
     modified_cosine = ModifiedCosine()
 
     with pytest.raises(AssertionError) as msg:
-        modified_cosine(norm_spectrum_1, norm_spectrum_2)
+        modified_cosine.pair(norm_spectrum_1, norm_spectrum_2)
 
     expected_message = "Precursor_mz missing. Apply 'add_precursor_mz' filter first."
     assert str(msg.value) == expected_message
@@ -37,7 +37,7 @@ def test_modified_cosine_with_mass_shift_5():
     norm_spectrum_1 = normalize_intensities(spectrum_1)
     norm_spectrum_2 = normalize_intensities(spectrum_2)
     modified_cosine = ModifiedCosine()
-    score, n_matches = modified_cosine(norm_spectrum_1, norm_spectrum_2)
+    score, n_matches = modified_cosine.pair(norm_spectrum_1, norm_spectrum_2)
 
     assert score == pytest.approx(0.081966, 0.0001), "Expected different cosine score."
     assert n_matches == 2, "Expected 2 matching peaks."
@@ -56,7 +56,7 @@ def test_modified_cosine_with_mass_shift_5_tolerance_2():
     norm_spectrum_1 = normalize_intensities(spectrum_1)
     norm_spectrum_2 = normalize_intensities(spectrum_2)
     modified_cosine = ModifiedCosine(tolerance=2.0)
-    score, n_matches = modified_cosine(norm_spectrum_1, norm_spectrum_2)
+    score, n_matches = modified_cosine.pair(norm_spectrum_1, norm_spectrum_2)
 
     assert score == pytest.approx(0.96788, 0.0001), "Expected different modified cosine score."
     assert n_matches == 6, "Expected 6 matching peaks."
@@ -75,8 +75,8 @@ def test_modified_cosine_order_of_input_spectrums():
     norm_spectrum_1 = normalize_intensities(spectrum_1)
     norm_spectrum_2 = normalize_intensities(spectrum_2)
     modified_cosine = ModifiedCosine(tolerance=2.0)
-    score_1_2, n_matches_1_2 = modified_cosine(norm_spectrum_1, norm_spectrum_2)
-    score_2_1, n_matches_2_1 = modified_cosine(norm_spectrum_2, norm_spectrum_1)
+    score_1_2, n_matches_1_2 = modified_cosine.pair(norm_spectrum_1, norm_spectrum_2)
+    score_2_1, n_matches_2_1 = modified_cosine.pair(norm_spectrum_2, norm_spectrum_1)
 
     assert score_1_2 == score_2_1, "Expected that the order of the arguments would not matter."
     assert n_matches_1_2 == n_matches_2_1, "Expected that the order of the arguments would not matter."